requirements.txt
python-dotenv
requests
flask
flask-cors
.env #with API Keys